backend/requirements.txt
fastapi
uvicorn
python-dotenv
boto3
httpx
pydantic
